Copied from Perforce
 Change: 1685595

+++ Code/CryEngine/RenderDll/XRenderD3D9/D3DOpenVR.h
@@ -21,7 +21,7 @@ public:
 	virtual void                      Shutdown() final;
 	virtual void                      OnResolutionChanged(int newWidth, int newHeight) final;
 	virtual void                      ReleaseBuffers() final {}
-	virtual void                      PrepareFrame(int frameId) final;
+	virtual void                      PrepareFrame(uint64_t frameId) final;
 	virtual void                      SubmitFrame() final;
 	virtual void                      OnPostPresent() final;
